reformatted with black

+++ mistreeplus/index/cantor.py
@@ -2,7 +2,9 @@ import numpy as np
 from typing import Union
 
 
-def cantor_pair(k1 : Union[int, np.ndarray], k2 : Union[int, np.ndarray]) -> Union[int, np.ndarray]:
+def cantor_pair(
+    k1: Union[int, np.ndarray], k2: Union[int, np.ndarray]
+) -> Union[int, np.ndarray]:
     """
     Uses the Cantor pairing function to construct a unique integer for two input
     integers.
@@ -19,15 +21,17 @@ def cantor_pair(k1 : Union[int, np.ndarray], k2 : Union[int, np.ndarray]) -> Uni
     pi : int/array
         Unique Cantor pair number.
     """
-    pi = 0.5*(k1 + k2 + 1.)*(k1 + k2) + k2
+    pi = 0.5 * (k1 + k2 + 1.0) * (k1 + k2) + k2
     if np.isscalar(k1) == True:
         pi = int(pi)
     else:
-        pi = pi.astype('int')
+        pi = pi.astype("int")
     return pi
 
 
-def uncantor_pair(pi : Union[int, np.ndarray]) -> tuple(Union[int, np.ndarray], Union[int, np.ndarray]):
+def uncantor_pair(
+    pi: Union[int, np.ndarray]
+) -> tuple(Union[int, np.ndarray], Union[int, np.ndarray]):
     """
     Reverses Cantor pairing function to determine the two input integers.
 
@@ -43,13 +47,13 @@ def uncantor_pair(pi : Union[int, np.ndarray]) -> tuple(Union[int, np.ndarray], 
     k2 : int/array
         Integer two.
     """
-    w = np.floor(0.5*(np.sqrt(8.*pi + 1.)-1.))
-    t = 0.5*(w**2. + w)
+    w = np.floor(0.5 * (np.sqrt(8.0 * pi + 1.0) - 1.0))
+    t = 0.5 * (w**2.0 + w)
     k2 = pi - t
     k1 = w - k2
     if np.isscalar(k1) == True:
         k1, k2 = int(k1), int(k2)
     else:
-        k1 = k1.astype('int')
-        k2 = k2.astype('int')
+        k1 = k1.astype("int")
+        k2 = k2.astype("int")
     return k1, k2
